BMD estimation for medrc or glsdrc models

Estimates the benchmark dose for fixed effects of medrc or glsdrc objects

BMD(object, respLev, interval = c("none", "delta", "fls", "tfls"),
    level = 0.95, bmd = c("additional", "extra"),
    background = 0.05, display=TRUE)

Arguments

object
an medrc object

respLev
a numeric vector containing the benchmark response levels

interval
character string specifying the type of confidence intervals to be supplied.

level
confidence level, default at 0.95

bmd
benchmark dose estimation (smallest dose resulting in a probability of an abnormal response)

background
probability of an abnormal response

display
logical. If TRUE results are displayed; otherwise they are not