Estimating marginal effective doses for medrc objects

ED estimates marginal effective doses (ECp/EDp/ICp) for given reponse levels, conditional on the estimated variance components.

EDmarg(object, respLev, interval = c("none", "delta", "fls", "tfls"),
  clevel = NULL, level = ifelse(!(interval == "none"), 0.95, NULL),
  reference = c("control", "upper"), type = c("relative", "absolute"),
  nGQ = 5, rfinterval = c(0, 1000), lref, uref, bound = TRUE,
  display = TRUE, logBase = NULL, ...)

Arguments

object
an object of class 'medrc'.

respLev
a numeric vector containing the response levels.

interval
character string specifying the type of confidence intervals to be supplied. The default is "none". Use "delta" for asymptotics-based confidence intervals (using the delta method and the t-distribution). Use "fls" for from logarithm scale based confidence intervals (in case the parameter in the model is log(ED50) as for the llogistic2) models. The only alternative for model-robust fits is using inverse regression.

clevel
character string specifying the curve id in case on estimates for a specific curve or compound is requested. By default estimates are shown for all curves.

level
numeric. The level for the confidence intervals. The default is 0.95.

reference
character string. Is the upper limit or the control level the reference?

type
character string. Whether the specified response levels are absolute or relative (default).

nGQ
integer. Specifies the number nof nodes for Gauss-Hermite quadrature.

rfinterval
numeric vector. Interval for root finding (uniroot) to search for ED values.

lref
numeric value specifying the lower limit to serve reference.

uref
numeric value specifying the upper limit to serve reference (eg. 100%).

bound
logical. If TRUE only ED values between 0 and 100% are allowed. FALSE is useful for hormesis models.

display
logical. If TRUE results are displayed. Otherwise they are not (useful in simulations).

logBase
numeric. The base of the logarithm in case logarithm transformed dose values are used.

...
additional arguments

Value

A matrix with two or more columns, containing the estimates and the corresponding estimated standard errors and possibly lower and upper confidence limits.