Asymptotic regression model

Providing the mean function and the corresponding self starter function for the asymptotic regression model.

meAR.2(x, d, e)

  meAR.3(x, c, d, e)

Arguments

x
numeric vector specifying the dose.

c
lower limit at at \(x=0\)

d
upper limit

e
\(e>0\) is determining the steepness of the increase as \(x\)

Details

The asymptotic regression model is a three-parameter model with mean function:

$$ f(x) = c + (d-c)(1-\exp(-x/e))$$

The parameter \(c\) is the lower limit (at \(x=0\)), the parameter \(d\) is the upper limit and the parameter \(e>0\) is determining the steepness of the increase as \(x\).

Value

A list of class drcMean, containing the mean function, the self starter function, the parameter names and other components such as derivatives and a function for calculating ED values.