Exponential decay model

Exponential decay model with or without a nonzero lower limit.

meEXD.2(x, d, e)

  meEXD.3(x, c, d, e)

Arguments

x
numeric dose vector.

c
lower limit

d
upper limit

e
steepness

Details

The exponential decay model is a three-parameter model with mean function:

$$f(x) = c + (d-c)(\exp(-x/e))$$

The parameter init is the upper limit (attained at \(x=0\)), the parameter plateau is the lower limit reached for x going to infinity and the parameter \(e>0\) is determining the steepness of the decay. The curve is monotonously decreasing in \(x\).

References

Organisation for Economic Co-operation and Development (OECD) (2006) Current approaches in the statistical analysis of ecotoxicity data: A guidance to application - annexes, Paris: OECD (p. 80).