Log-normal dose-response model

LN.X and the accompanying convenience functions provide a general framework for specifying the mean function of the decreasing or incresing log-normal dose-response model.

meLN.2(x, b, e)

meLN.3(x, b, d, e)

meLN.4(x, b, c, d, e)

meLLN.4(x, b, c, d, e)

Arguments

x
numeric dose vector

b
steepness

e
ED50

d
upper limit

c
lower limit

Details

For the case where log(ED50), denoted \(e\) in the equation below, is a parameter in the model, the mean function is: $$f(x) = c + (d-c)(\Phi(b(\log(x)-e)))$$ and the mean function is: $$f(x) = c + (d-c)(\Phi(b(\log(x)-\log(e))))$$ in case ED50, which is also denoted \(e\), is a parameter in the model. If the former model is fitted any estimated ED values will need to be back-transformed subsequently in order to obtain effective doses on the original scale. The mean functions above yield the same models as those described by Bruce and Versteeg (1992), but using a different parameterisation (among other things the natural logarithm is used). For the case \(c=0\) and \(d=1\), the log-normal model reduces the classic probit model (Finney, 1971) with log dose as explanatory variable (mostly used for quantal data). This special case is available through the convenience function meLN.2. The case \(c=0\) is available as the function meLN.3. The full four-parameter model is available through meLN.4.

References

Finney, D. J. (1971) Probit analysis, London: Cambridge University Press. Bruce, R. D. and Versteeg, D. J. (1992) A statistical procedure for modeling continuous toxicity data, Environ. Toxicol. Chem., 11, 1485--1494.