Four-parameter Weibull functions

'W1.X' and 'W2.X' provide the X-parameter Weibull functions.

meW1.2(x, b, e)

meW1.3(x, b, d, e)

meW1.4(x, b, c, d, e)

meW2.2(x, b, e)

meW2.3(x, b, d, e)

meW2.4(x, b, c, d, e)

Arguments

x
numeric dose vector.

b
steepness

e
ED50

d
upper limit

c
lower limit

Details

As pointed out in Seber and Wild (1989), there exist two different parameterisations of the Weibull model. They do not yield the same fitted curve for a given dataset. The four-parameter Weibull type 1 model ('weibull1') is $$ f(x) = c + (d-c) \exp(-\exp(b(\log(x)-\log(e)))).$$ Thw four-parameter Weibull type 2 model ('weibull2') is $$ f(x) = c + (d-c) (1 - \exp(-\exp(b(\log(x)-\log(e))))).$$ Both four-parameter model functions are asymmetric with inflection point at the dose equal \(e\).

References

Seber, G. A. F. and Wild, C. J (1989) Nonlinear Regression, New York: Wiley \& Sons (pp. 330--331). Ritz, C (2009) Towards a unified approach to dose-response modeling in ecotoxicology To appear in Environ Toxicol Chem.