medrc methods

Methods for objects of class medrc

# S3 method for medrc
df.residual(object, ...)
# S3 method for medrc
predict(object, ..., newdata = NULL,
                        level = NULL,
                        type=c("conditional", "marginal"),
                        nGQ=5)
# S3 method for medrc
residuals(object, ...)
# S3 method for medrc
vcov(object, ...)
# S3 method for medrc
summary(object, ...)
# S3 method for medrc
ranef(object, ...)
# S3 method for medrc
VarCorr(x, sigma = 1, ...)
# S3 method for medrc
AIC(object, ..., k = 2)
# S3 method for medrc
BIC(object, ...)
# S3 method for medrc
logLik(object, REML = FALSE, ...)
# S3 method for medrc
print(x, ..., digits = max(3, getOption("digits") - 3))

Arguments

object
An object of class medrc

x
An object of class medrc

newdata
an optional data frame to be used for obtaining the predictions. All variables used in the fixed and random effects models, as well as the grouping factors, must be present in the data frame. If missing, the fitted values are returned.

level
an optional integer vector giving the level(s) of grouping to be used in obtaining the predictions. Level values increase from outermost to innermost grouping, with level zero corresponding to the population predictions. Defaults to the highest or innermost level of grouping.

type
character, either "conditional" or "marginal" for predictions conditional on the random effects or marginal predictions (conditional on the estimated variance components).

nGQ
integer. Number of nodes for Gauss-Hermite quadrature, when type = "marginal".

k
numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.

REML
an optional logical value. If TRUE the restricted log-likelihood is returned, else, if FALSE, the log-likelihood is returned. Defaults to FALSE.

sigma
an optional numeric value used as a multiplier for the standard deviations. Default is 1.

digits
minimal number of significant digits

...
further arguments