Mixed effects dose-response curves

Implements drc nonlinear functions into the nlme framework for mixed effects dose-response modelling.

medrm(form, curveid = NULL, data, fct, random, correlation = NULL,
  weights = NULL, control = NULL, start = NULL, REML = FALSE)

Arguments

form
Formula describing the dose-response relationship

curveid
Formula with parameter names on the left hand side (divided by +) and a column name in data, denoting a factor, to estimate separate parameters per factor-level. If NULL only fixed effects for a single curve will be estimated.

data
a data.frame object

fct
a function of the drc package

random
a list or one-sided formula describing the random effects

correlation
additional corClasses object

weights
additional varClasses object

control
list with nlme control arguments

start
optional list with initial values for the fixed components. If NULL the initial values will be found automatically.

REML
logical value. If TRUE the model is fit by maximizing the restricted log-likelihood, if FALSE log-likelihood is maximized.

Value

An object of class medrc

Details

An application of medrm is shown on the help pages of data broccoli. EDx and selectivity indices can be calculated with functions ED and EDcomp. Model-averaged ED can be computed by function mmaED. Marginal EDx and corresponding function are implemented in functions EDmarg, using numerical integration conditional on the etimated variance components.

See also

drm, nlme